More object libraries: constants, stringx, stats Broke a needless dependency chain by extracting some string functions from sm/misc/utils.* and moving them to common/stdx_string.*
