+++ tiledb/common/CMakeLists.txt
@@ -39,14 +39,13 @@ get_gathered(COMMON_SOURCES)
 set(TILEDB_COMMON_SOURCES ${COMMON_SOURCES} PARENT_SCOPE)
 
 #
-# Object library for other units to depend upon
+# `baseline` object library
 #
 add_library(baseline OBJECT
     logger.cc status.cc governor/governor.cc heap_profiler.cc heap_memory.cc
 )
 find_package(Spdlog_EP REQUIRED)
 target_link_libraries(baseline PUBLIC spdlog::spdlog)
-
 #
 # Test-compile of object library ensures link-completeness
 #
@@ -54,4 +53,19 @@ add_executable(compile_baseline EXCLUDE_FROM_ALL)
 target_link_libraries(compile_baseline PRIVATE baseline)
 target_sources(compile_baseline PRIVATE
     test/compile_baseline_main.cc $<TARGET_OBJECTS:baseline>
+)
+
+#
+# `stringx` object library
+#
+add_library(stringx OBJECT
+    stdx_string.cc
+)
+#
+# Test-compile of object library ensures link-completeness
+#
+add_executable(compile_stringx EXCLUDE_FROM_ALL)
+target_link_libraries(compile_stringx PRIVATE stringx)
+target_sources(compile_stringx PRIVATE
+    test/compile_stringx_main.cc $<TARGET_OBJECTS:stringx>
 )
