+++ tiledb/sm/filesystem/win.cc
@@ -34,6 +34,7 @@
 #include "tiledb/sm/filesystem/win.h"
 #include "tiledb/common/heap_memory.h"
 #include "tiledb/common/logger.h"
+#include "tiledb/common/stdx_string.h"
 #include "tiledb/sm/misc/constants.h"
 #include "tiledb/sm/misc/utils.h"
