+++ tiledb/sm/misc/uri.cc
@@ -32,6 +32,7 @@
 
 #include "tiledb/sm/misc/uri.h"
 #include "tiledb/common/logger.h"
+#include "tiledb/common/stdx_string.h"
 #include "tiledb/sm/filesystem/vfs.h"
 #include "tiledb/sm/misc/utils.h"
