+++ tiledb/sm/misc/utils.h
@@ -160,23 +160,6 @@ bool is_int(const std::string& str);
 /** Returns `true` if the input string is an unsigned integer. */
 bool is_uint(const std::string& str);
 
-/**
- * Checks if a string starts with a certain prefix.
- *
- * @param value The base string.
- * @param prefix The prefix string to be tested.
- * @return *true* if *value* starts with the *prefix*, and *false* otherwise.
- */
-bool starts_with(const std::string& value, const std::string& prefix);
-
-/**
- * Checks if a string ends with a certain suffix.
- * @param value The base string.
- * @param suffix The suffix to be tested.
- * @return *true* if *value* ends with the *suffix*, and *false* otherwise.
- */
-bool ends_with(const std::string& value, const std::string& suffix);
-
 /** Converts the input value to string. */
 template <class T>
 std::string to_str(const T& value);
