+++ tiledb/sm/stats/stats.cc
@@ -31,6 +31,7 @@
  */
 
 #include "tiledb/sm/stats/stats.h"
+#include "tiledb/common/stdx_string.h"
 #include "tiledb/sm/misc/utils.h"
 
 #include <algorithm>
